Decorators should parse now
A few failing cases to fix, plus refactor of  class parsing code.

This implementation differs from Babylon implementation, because I avoid this to be restricted to the 'parseDecorator'  production. Instead it is designed to easely implement future decorator spec changes.

'decoratorMemberExpression' and 'decoratorCallExpression' productions are both part of their methods without the 'decorator' prefix.