refactor(lexer) tweaked comment scanning
I may have done some stupidity with the single line comment, but when I ran the same tests after each other I could never go higher than 22ms. Without this optimization it wasn't stable, and I could reach 80ms and above.

Mostly the single line comment are followed by a newline, so maybe an overkill to check if the 4th bit set is set or not. But I guess it's worth a try